Bug fix in ColdEntityWorker
We schedule the next run if there are more requests than allowed in a single batch. Previously, we used filteredRequests.size > batchSize to decide whether a next run is required or not. But filteredRequests may be low despite there are still a lot of requests in the queue. This PR fixes the bug by using requestSize > batchSize instead.

This PR also changes the return type of pullRequests as we don’t use its return value.

Testing done:
1. Added unit tests for the changed code.